[Feature] Issue #42 Persistent AUTO_INCREMENT value for InnoDB table
InnoDB auto_increment value is not persistent and can lead to some
values reused, as described by official bug #199.

This patch fix the problem by writing the last used AUTO_INCREMENT
to the root page of the clustered index, using the PAGE_MAX_TRX_ID
field, which is unused for root page.

Two InnoDB variables are introduced to control this persistent behaviour:
- innodb_autoinc_persistent
  This persistent behaviour is enabled or not.

- innodb_autoinc_persistent_interval
  The interval of persist max auto_increment value.