[bugfix] Issue: #64 replication regression with RBR and partition tables
This patch is ported from MySQL upstream (commit id: 9414c7).

Bug#25687813 REPLICATION REGRESSION WITH RBR AND PARTITIONED TABLES

PROBLEM
-------

While applying update the slave is trying to initialize all partitions
before reading from rnd_pos() call. This is done for each row update
because of which the performance is getting effected.

FIX
---

Initialize only the partition on which rnd_pos() is called.